refactor: move DeleteStockUiState to dedicated state package
Moves the DeleteStockUiState data class out of the ViewModel file into the new ui/state package for better separation of concerns. Creates the DeleteStockUiState.kt file within the ui/state directory and removes the inline definition from DeleteStockViewModel. All relevant imports in the ViewModel are updated to reflect the new file location.
